Only ts files at the src level are targeted
Create a project. Create a widget
VS code gives the error:

"Experimental support for decorators is a feature that is subject to change in a future release. Set the 'experimentalDecorators' option to remove this warning."

cmd+shift+p "TypeScript: Go to project configuration" says the file is not part of the project.

Extend the glob to include folders below